PubSub system
This is an implementation of a distributed publisher-subscriber system
using Java Remote Method Invocation (RMI). It uses multi-threading with
thread-per-request and does not block for send or receive.  Both the
server and the client have a command line interface for publishing
events, advertising events, subscribing to topics or keywords, and
viewing past events received.  The system supports both Topic-based and
Content-based filtering of events, where the content filtering is based
off of keyword arguments.  They keywords can be defined by the Topic or
they can be defined by the Event, and if this keyword matches the
subscribed keyword of a client then the client will receive the message.
Using the PubSub system, there is also a small implementation of a stock
market, where buyers and sellers communicate to one another through
Pub-Sub and then use RMI to finalize buys or sells.